dsc-drivers: update ionic to 25.01.1-001 (#52)
* dsc-drivers: update ionic to 24.12.4-001

Several updates including:
  - support for new Salina device RDMA driver
  - new kernel_compat.sh feature for better auto-configure of various distribution kernels
  - better ethtool support for 50, 200, and 400 Gbe QSFP modules

Internal patch list:
    ionic_rdma: queue group oriented dual-udma
    ionic: MOFED aux device compatibility
    ionic: updates for Linux kernel v6.13
    ionic: set pci max seg size to 1G.
    ionic: use ee->offset when returning sprom data
    ionic: add new qsfp type support
    ionic: remove unsupported rx rule related code
    ionic: kcompat tweaks for RHEL 9.5
    ionic: Fix the ethtool outputs issues
    PS-13181: Configure CMB mapping for WQE size 128 B
    ionic_rdma: autocompat bitfield members
    tools: autocompat: change outdir to KPATH if taking module path
    tools: autocompat: Take kbuild path from $(srctree) for 6.13 onwards
    tools: autocompat: use custom filechk to store temp autocompat.h
    eth: add autocompat support for OL related failures
    ionic: add speed defines for 200G and 400G
    build: tools: add auto compat functionality for compilation on diff flavors
    ionic: override expdb feature negotiation
    ionic_rdma: expdb support 2 mappings
    Add autocompat for eth and mnic driver
    build: tools: add auto compat functionality for compilation on diff flavors
    ionic: export ionic_api_adminq_post_wait API
    PS-13017: add CMB discovery feature bits
    ionic: add ASIC_TYPE defines to ionic_if
    ionic: fix kcompat for oracle kernel 8.8
    ionic: fix kcompat for oracle kernel 8.9
    create auxiliary device from device probe
    Register auxbus driver from ionic rdma driver
    add enable_rdma devlink parameter support
    Register aux device from ionic pcie driver.
    ionic: Prevent phc worker from running/restarting during shutdown
    ionic: Fix compilation issue to to mismatched format specifier
    ionic: Fix compilation issue on linux mainline when !CONFIG_NET_DEVLINK
    ionic: no double destroy workqueue
    ionic: Add kcompat(ish) support for sysfs_emit
    ionic: Translate IONIC_RC_ENOSUPP to EOPNOTSUPP
    ionic: Add net sysfs support for setting the DSC's int_mnic address
    ionic: Kcompat for v6.11 ethtool timestamp API changes
    ionic: Fix netpoll case where budget == 0
    ionic: Use VLAN_ETH_HLEN when possible


* dsc-drivers: update ionic to 25.01.1-001

Includes:
    ionic: fix debugfs q stats pointer (#455)
    ionic_rdma: clear CMB requested region in eth driver (AI-367)
    ionic: change devlink registration sequence
    ionic: rework compat handling of devlink param ENABLE_RDMA


---------

+++ drivers/linux/eth/ionic/ionic_fw.c
@@ -98,7 +98,7 @@ int ionic_firmware_update(struct ionic_lif *lif, const struct firmware *fw)
 
 	netdev_dbg(netdev,
 		   "downloading firmware - size %d part_sz %d nparts %lu\n",
-		   (int)fw->size, buf_sz, DIV_ROUND_UP(fw->size, buf_sz));
+		   (int)fw->size, buf_sz, (unsigned long)DIV_ROUND_UP(fw->size, buf_sz));
 
 	devlink_flash_update_status_notify(dl, "Downloading", NULL, 0, fw->size);
 	offset = 0;
@@ -115,7 +115,7 @@ int ionic_firmware_update(struct ionic_lif *lif, const struct firmware *fw)
 		if (err) {
 			netdev_err(netdev,
 				   "download failed offset 0x%x addr 0x%lx len 0x%x\n",
-				   offset, offsetof(union ionic_dev_cmd_regs, data),
+				   offset, (unsigned long)offsetof(union ionic_dev_cmd_regs, data),
 				   copy_sz);
 			goto err_out;
 		}
